new version
aaaa

+++ GranBlueFantasy_And_OtherCygame_Support_Dll/Feat/Feat_Vajra.cs
@@ -165,7 +165,7 @@ namespace GBF.feat.Feat_Vajra
                         }
                     }
                     // 跳过1218检查，直接应用其他抗性 / Skip 1218 check, directly apply other resistances / 1218チェックをスキップ、直接他の抵抗を適用
-                    damage = Element.GetResistDamage((int)damage, target.Evalue(e.source.aliasRef), resistBonus);
+                    damage = Element.GetResistDamage(damage, target.Evalue(e.source.aliasRef), resistBonus);
                 }
             }
